Calendar: Various optimizations
